refactor: 조건에 해당하는 List<Direction>을 반환해주는 메서드를 Drections 클래스로 분리
